job descriptions: give the same order for all jobs descriptions
