+++ itou/templates/siaes/card.html
@@ -5,81 +5,79 @@
 {% comment %} {% block nb_columns %}10{% endblock %} {% endcomment %}
 
 {% block content %}
-    {% with siae.job_description_through.all as jobs_descriptions %}
-        <div class="card c-card  c-card--emploi p-3">
-            <div class="card-body">
-                <div class="pb-3">
-                    <h1 class="h2 mb-2">
-                        <b><abbr title="{{ siae.get_kind_display }}">{{ siae.kind }}</abbr></b>
-                        -
-                        <a href="{{ siae.get_card_url }}?back_url={{ request.get_full_path|urlencode }}" class="matomo-event" data-matomo-category="candidature" data-matomo-action="clic" data-matomo-option="clic-structure">
-                            {{ siae.display_name }}
-                        </a>
-                        {# Display non-user-edited name too. #}
-                        {% comment %} {% if siae.brand %}<small class="text-muted">({{ siae.name|title }})</small>{% endif %} {% endcomment %}
-                    </h1>
-                    <adresse class="card-subtitle font-size-hat font-weight-bold text-muted">{{ siae.address_on_one_line }}</adresse>
-                </div>
-
-                {% if siae.email or siae.phone or siae.website %}
-                    <div class="py-3">
-                        <h6>Contacts</h6>
-                        {% if siae.email %}
-                            <p>
-                                {% include "includes/icon.html" with icon="mail" title="E-mail" %}
-                                <a href="mailto:{{ siae.email }}">{{ siae.email }}</a>
-                            </p>
-                        {% endif %}
-
-                        {% if siae.phone %}
-                            <p>
-                                {% include "includes/icon.html" with icon="phone" title="Téléphone" %}
-                                <a href="tel:{{ siae.phone }}">{{ siae.phone|format_phone }}</a>
-                            </p>
-                        {% endif %}
-
-                        {% if siae.website %}
-                            <p>
-                                {% include "includes/icon.html" with icon="external-link" title="Site Internet" %}
-                                <a href="{{ siae.website }}" rel="noopener" target="_blank">{{ siae.website }}</a>
-                            </p>
-                        {% endif %}
-
-                    </div>
-                {% endif %}
-                {% if siae.description %}
-                    <div class="border-top pt-3">
-                        <h6 class="mb-1">Descriptif</h6>
-                        {{ siae.description|linebreaks }}
-                    </div>
-                {% endif %}
-                {% if jobs_descriptions %}
-                    {% include "siaes/includes/_list_siae_jobs.html" with siae=siae jobs_descriptions=jobs_descriptions %}
-                {% endif %}
-                {% if siae.count_active_job_descriptions > 0 and not siae.block_job_applications %}
-                    <div class="border-top py-3">
-                        {% include "siaes/includes/_list_siae_actives_jobs.html" with jobs_descriptions=jobs_descriptions siae=siae %}
-                    </div>
-                {% endif %}
-                {% if not siae.block_job_applications %}
-                    <div class="border-top pt-4">
-                        <p class="text-with-btn">
-                            <span class="h4">Cette structure vous interesse ?</span>
-                            <a class="btn btn-primary float-right"
-                                href="{% url 'apply:start' siae_pk=siae.pk %}{% if back_url %}?back_url={{ back_url }}{% endif %}"
-                                aria-label="Postuler auprès de l'employeur solidaire {{ siae.display_name }}">
-                                Proposez votre candidature
-                            </a>
-                        </p>
-                    </div>
-                {% endif %}
+    <div class="card c-card  c-card--emploi p-3">
+        <div class="card-body">
+            <div class="pb-3">
+                <h1 class="h2 mb-2">
+                    <b><abbr title="{{ siae.get_kind_display }}">{{ siae.kind }}</abbr></b>
+                    -
+                    <a href="{{ siae.get_card_url }}?back_url={{ request.get_full_path|urlencode }}" class="matomo-event" data-matomo-category="candidature" data-matomo-action="clic" data-matomo-option="clic-structure">
+                        {{ siae.display_name }}
+                    </a>
+                    {# Display non-user-edited name too. #}
+                    {% comment %} {% if siae.brand %}<small class="text-muted">({{ siae.name|title }})</small>{% endif %} {% endcomment %}
+                </h1>
+                <adresse class="card-subtitle font-size-hat font-weight-bold text-muted">{{ siae.address_on_one_line }}</adresse>
             </div>
-        </div>
 
-        {% if back_url %}
-            <p class="mt-4">
-                <a href="{{ back_url }}">Retour</a>
-            </p>
-        {% endif %}
-    {% endwith %}
+            {% if siae.email or siae.phone or siae.website %}
+                <div class="py-3">
+                    <h6>Contacts</h6>
+                    {% if siae.email %}
+                        <p>
+                            {% include "includes/icon.html" with icon="mail" title="E-mail" %}
+                            <a href="mailto:{{ siae.email }}">{{ siae.email }}</a>
+                        </p>
+                    {% endif %}
+
+                    {% if siae.phone %}
+                        <p>
+                            {% include "includes/icon.html" with icon="phone" title="Téléphone" %}
+                            <a href="tel:{{ siae.phone }}">{{ siae.phone|format_phone }}</a>
+                        </p>
+                    {% endif %}
+
+                    {% if siae.website %}
+                        <p>
+                            {% include "includes/icon.html" with icon="external-link" title="Site Internet" %}
+                            <a href="{{ siae.website }}" rel="noopener" target="_blank">{{ siae.website }}</a>
+                        </p>
+                    {% endif %}
+
+                </div>
+            {% endif %}
+            {% if siae.description %}
+                <div class="border-top pt-3">
+                    <h6 class="mb-1">Descriptif</h6>
+                    {{ siae.description|linebreaks }}
+                </div>
+            {% endif %}
+            {% if jobs_descriptions %}
+                {% include "siaes/includes/_list_siae_jobs.html" with siae=siae jobs_descriptions=jobs_descriptions %}
+            {% endif %}
+            {% if siae.count_active_job_descriptions > 0 and not siae.block_job_applications %}
+                <div class="border-top py-3">
+                    {% include "siaes/includes/_list_siae_actives_jobs.html" with jobs_descriptions=jobs_descriptions siae=siae %}
+                </div>
+            {% endif %}
+            {% if not siae.block_job_applications %}
+                <div class="border-top pt-4">
+                    <p class="text-with-btn">
+                        <span class="h4">Cette structure vous interesse ?</span>
+                        <a class="btn btn-primary float-right"
+                            href="{% url 'apply:start' siae_pk=siae.pk %}{% if back_url %}?back_url={{ back_url }}{% endif %}"
+                            aria-label="Postuler auprès de l'employeur solidaire {{ siae.display_name }}">
+                            Proposez votre candidature
+                        </a>
+                    </p>
+                </div>
+            {% endif %}
+        </div>
+    </div>
+
+    {% if back_url %}
+        <p class="mt-4">
+            <a href="{{ back_url }}">Retour</a>
+        </p>
+    {% endif %}
 {% endblock %}
